+++ crypto/encrypt.js
