+++ customer.js
@@ -1,11 +1,11 @@
 // This script runs on serverside
 
 //dependencies required for the app
-var connection = require('../Pw-manager-nodejs/database');
-var encrypt1 = require('../Pw-manager-nodejs/encrypt');
+var connection = require('../Pw-manager-nodejs/Database/database');
+var encrypt1 = require('../Pw-manager-nodejs/crypto/encrypt');
 const dateLib = require('date-and-time');
 var User = require('../Pw-manager-nodejs/user');
-var encrypt1 = require('../Pw-manager-nodejs/encrypt');
+var encrypt1 = require('../Pw-manager-nodejs/crypto/encrypt');
 const bcrypt = require('bcrypt');
 var escape = require('lodash.escape');
 
@@ -13,20 +13,20 @@ var escape = require('lodash.escape');
 
 async function signUp(req) {
 
-    var pw = req.body.pw;
-    var pw1 = req.body.pw1;
+    let pw = req.body.pw;
+    let pw1 = req.body.pw1;
 
     if (pw !== pw1) {
         return "pw missmatch";
     }
 
-    var hashedPw = encrypt1.hashPw(pw);
+    let hashedPw = encrypt1.hashPw(pw);
     if (hashedPw === null) {
         return "error: Pw hash problem!";
     }
     // TODO: Autoincrement in mysql
-    var id = helper.getRandomInt(1, 10000).toString();
-    var user = new User(id, req.body.username, req.body.surname, req.body.lastname, hashedPw, null);
+    let id = helper.getRandomInt(1, 10000).toString();
+    let user = new User(id, req.body.username, req.body.surname, req.body.lastname, hashedPw, null);
 
     if (pw === null || pw1 === null) {
         return "error: Pw not found!";
@@ -37,13 +37,13 @@ async function signUp(req) {
     }
 
     try {
-        var userExists = await connection.getUserExists(user.username);
+        let userExists = await connection.getUserExists(user.username);
 
         if(userExists){
             return "User already exists!";
         }
 
-        var tempDate = new Date();
+        let tempDate = new Date();
         tempDate = dateLib.format(tempDate, 'YYYY-MM-DD');
     
         await connection.addUser(user);
@@ -57,7 +57,12 @@ async function signUp(req) {
 };
 
 
-
+/** signin user and store to session
+ * 
+ * @param {*} req 
+ * @param {*} res 
+ * @returns 
+ */
 async function signIn(req, res) {
 
     if (req.session.loggedIn) {
@@ -84,7 +89,12 @@ async function signIn(req, res) {
 };
 
 
-
+/** Logout the current user from the session
+ * 
+ * @param {*} req 
+ * @param {*} res 
+ * @returns 
+ */
 function logout(req, res) {
     req.session.loggedIn = false;
     req.session.username = "";
@@ -98,7 +108,7 @@ function logout(req, res) {
 
 function setUserToSession(req, res, user) {
     console.log("in setUserToSession");
-    var tempEncryptPW = encrypt1.hashPw(req.body.pw);
+    let tempEncryptPW = encrypt1.hashPw(req.body.pw);
 
     if (bcrypt.compare(tempEncryptPW, user.pw)) {
         req.session.loggedIn = true;
@@ -114,7 +124,12 @@ function setUserToSession(req, res, user) {
 }
 
 
-
+/** Gets the current user from the session
+ * 
+ * @param {*} req 
+ * @param {*} res 
+ * @returns User object
+ */
 function getUserFromSession(req, res) {
     return new User(req.session.id, req.session.username, req.session.surname, req.session.lastname, null, req.session.loggedIn);
 };
