+++ helper.js
@@ -3,7 +3,7 @@
 const dateLib = require('date-and-time');
 
 function getCurrentDate() {
-    var tempDate = new Date();
+    let tempDate = new Date();
     return dateLib.format(tempDate, 'YYYY-MM-DD');
 }
